pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["y"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flake8 = "^7.1.1"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
validators = "^0.33.0"
requests = "^2.32.3"
bs4 = "^0.0.2"
data = "^0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
